@@ -1,65 +1,43 @@
-buildscript {
-    ext {
-        kotlinVersion = '1.5.31'
-        springBootVersion = '2.1.0.RELEASE'
-    }
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}")
-        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:${kotlinVersion}")
-        classpath("org.jetbrains.kotlin:kotlin-allopen:${kotlinVersion}")
-    }
-}
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
-apply plugin: 'kotlin'
-apply plugin: 'kotlin-spring'
-apply plugin: 'eclipse'
-apply plugin: 'org.springframework.boot'
-apply plugin: 'io.spring.dependency-management'
+plugins {
+	id 'org.springframework.boot' version '3.0.6'
+	id 'io.spring.dependency-management' version '1.1.0'
+	id 'org.jetbrains.kotlin.jvm' version '1.7.22'
+	id 'org.jetbrains.kotlin.plugin.spring' version '1.7.22'
+}
 
 group = 'io.pivotal.metric_registrar.examples'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 1.8
-compileKotlin {
-    kotlinOptions {
-        freeCompilerArgs = ["-Xjsr305=strict"]
-        jvmTarget = "1.8"
-    }
-}
-compileTestKotlin {
-    kotlinOptions {
-        freeCompilerArgs = ["-Xjsr305=strict"]
-        jvmTarget = "1.8"
-    }
-}
+sourceCompatibility = '17'
 
 repositories {
     mavenCentral()
 }
 
-test {
-    useJUnitPlatform()
-}
-
 dependencies {
     implementation(
-        'org.springframework.boot:spring-boot-starter-actuator',
+        'org.springframework.boot:spring-boot-starter',
         'org.springframework.boot:spring-boot-starter-web',
-
-        'io.micrometer:micrometer-registry-prometheus:latest.release',
-
-        'com.fasterxml.jackson.module:jackson-module-kotlin',
-        'org.jetbrains.kotlin:kotlin-stdlib-jdk8',
         'org.jetbrains.kotlin:kotlin-reflect',
+
+		'io.micrometer:micrometer-registry-prometheus'
     )
 
+    testImplementation (
+		'org.springframework.boot:spring-boot-starter-test',
 
-    testImplementation(
-        'org.springframework.boot:spring-boot-starter-test',
-        'org.junit.jupiter:junit-jupiter-api:5.3.1',
-        'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0'
-    )
-    testRuntimeOnly('org.junit.jupiter:junit-jupiter-engine')
+		'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0'
+	)
+}
+
+tasks.withType(KotlinCompile) {
+	kotlinOptions {
+		freeCompilerArgs = ['-Xjsr305=strict']
+		jvmTarget = '17'
+	}
+}
+
+tasks.named('test') {
+	useJUnitPlatform()
 }
